optional commandline args and further cleanup

+++ heic-conv.py
@@ -1,9 +1,24 @@
-import os, subprocess
+'''A pythons script for batch converting heic files.'''
+
+
+import argparse
+import os
+import subprocess
 
 
 CMD = "heif-convert"
 QUALITY = "-q"
 QUALITY_ARG = "99"
+AUX_FILE_SUFFIX = "-urn:com:apple:photo:2020:aux:hdrgainmap"
+
+
+def parse_args() -> argparse.Namespace:
+    '''Get the args from the command line'''
+    parser = argparse.ArgumentParser()
+    parser.add_argument("--daux", action="store_true", help="Delete aux files")
+    parser.add_argument("--dorig", action="store_true", help="Delete original files")
+    args = parser.parse_args()
+    return args
 
 
 def find_all_heics() -> None:
@@ -13,9 +28,8 @@ def find_all_heics() -> None:
     return heics
 
 
-def convert_and_delete(file_name: str) -> None:
+def convert_and_delete(file_name: str, delete_aux: bool = False, delete_orig:bool = False) -> None:
     '''On the original file name, call heif-convert, and delete the aux and original files'''
-    
     base_file = file_name[:-5]
     original_file_extension = file_name[-5:]
 
@@ -28,19 +42,22 @@ def convert_and_delete(file_name: str) -> None:
     string_output = output.decode("utf-8")
     print(string_output)
 
-    # delete the auxilary file
-    aux_file_name = base_file + "-urn:com:apple:photo:2020:aux:hdrgainmap" + ".JPG"
-    os.remove(aux_file_name)
-    print(f"Deleted aux file: {aux_file_name}")
+    if delete_aux:
+        # delete the auxilary file
+        aux_file_name = base_file + AUX_FILE_SUFFIX + ".JPG"
+        os.remove(aux_file_name)
+        print(f"Deleted aux file: {aux_file_name}")
 
-    # delete the original file
-    orig_file_name = base_file + original_file_extension
-    os.remove(orig_file_name)
-    print(f"Deleted original file: {orig_file_name}")
-
-    print("\n")
+    if delete_orig:
+        # delete the original file
+        orig_file_name = base_file + original_file_extension
+        os.remove(orig_file_name)
+        print(f"Deleted original file: {orig_file_name}\n")
 
 if __name__ == "__main__":
+    args = parse_args()
     heics = find_all_heics()
+    print(f'Found {len(heics)} heic files in this directory.')
+
     for file in heics:
-        convert_and_delete(file)
+        convert_and_delete(file, args.daux, args.dorig)
